Removed or moved code from opt plot file
