+++ openmdao/visualization/realtime_opt_plot/realtime_plot_class.py
@@ -1,12 +1,16 @@
+"""
+  The abstract class for the two forms of real time plot which are for AnalysisDriver and optimizers.
+  Put in a separate class to avoid circular imports
+"""
 class _RealTimePlot(object):
     def __init__(
-        self, case_tracker, callback_period, doc, pid_of_calling_script   # TODO why isn't callback_period used?
+        self, case_tracker, callback_period, doc, pid_of_calling_script
     ):
         """
         Construct and initialize _RealTimeOptPlot instance.
         """
-        # self._case_recorder_filename = case_recorder_filename
         self._case_tracker = case_tracker
         self._pid_of_calling_script = pid_of_calling_script
         self._doc = doc
+        self._callback_period = callback_period
